requirements.txt
Pillow
streamlit
langchain
langchain-huggingface
langchain-ollama
langchain-openai
python-dotenv
